Add Categories Array
In this code, a new file "categories.h" was created, where the array
categories was declared, containing the names of the menu categories.
The variable num_categories is also declared, which stores the number
of categories. The "categories.h" file is included using the
 #include "categories.h" directive. The for loop in the main file
"main_menu.c" displays the names of the categories using the
categories array and the num_categories variable.

Now the program can work with a dynamic list of categories,
which makes it easier to add or remove categories in the future.