Added checks and functional for choosing a date
If the month is odd, it sets the `month_end_day` variable to 30.
This is because months with odd numbers (January, March, May, etc.)
typically have 31 days, except for February.

If the month is February it sets the `month_end_day` variable to 28.
This is because February usually has 28 days, except in leap years.

If the year is a leap year (2024, 2028, 2032), and the month is February
the last day of the month is set to 29.

If none of the previous conditions are met.
It sets the `month_end_day` variable to 31. This covers the months with
even numbers (April, June, August, etc.) which generally have 31 days.

You can choose a day between the current day
and the last day of the month.

If the chosen date is the current day and the current hour is 20
or later, a message is displayed indicating that no more tickets are
available for the day.