The today_time_check function is designed to check if the current time has exceeded the specified session time for a movie on the selected date.
It compares the day of the dt struct with the selected date
(date_session) to determine if it is the same day.

If the day matches, it checks if the current hour (dt->hour) is greater
than or equal to the session time (session_time).

If the current hour is greater than or equal to the session time
it means that the session has expired, and a message is displayed
to inform the user.

The purchasing_function is called again to allow the user to select
another session for ticket purchase.