Added function popcorn_store
The `popcorn_store` function presents a menu to the user for selecting
different options related to purchasing popcorn and drinks.

The menu options are presented to the user, including options for
popcorn and drinks combinations, as well as individual options for only
popcorn, only drinks, or neither.

Variables `money_spended` , and the `popcorn_menu` is set to number
depends on chosen commodity. Then, the `check_duplicate_tickets`
function is called to ensure there are no duplicate tickets.