The `purchase_accept` function is responsible for processing the purchase of a movie ticket.
Calls the `check_duplicate_tickets` function, which checks for duplicate
tickets for a given session.
This is important to prevent repeated purchases of tickets for the same
session.

Displays information about the movie for which the user is buying
a ticket.

If the user confirms the purchase (`purchase_confirm == 1`),
the data about the active session is saved in the `active_session`
structure, the total amount of money and the number of purchased tickets
are additionally calculated and saved.

If the user cancels the purchase (`purchase_confirm == 0`),
a corresponding message is displayed.And returns the user to the movie
catalog