Added constant 'const char* history_filename="History.txt";'
The added function, `check_duplicate_tickets`, checks for duplicate
tickets for the current movie session.
By reading data from the history file.If the file does not exist,
it creates a new file. After successfully creating the file,
the function immediately calls `purchase_accept`
(because there is no data to compare) to continue the ticketing process.

The function reads the file line by line and extracts ticket data from
each line. The read data from each line is compared with the data of the
current ticket. If a duplicate ticket is found
(compared to the following data:
film_id, date, time, row, place, localization) a message about the found
duplicate is displayed, and the `purchasing_function` is called
to change the ticket data.