Changing the declaration of the history_filename constant
Added constant #define CODE_LENGTH 10

The added `ticket_creator` function is responsible for creating a ticket
for the active movie session. It calls the `generate_unique_code`
function, which generates a unique code for the ticket. Ticket data is
written to the file '"ticket%d.txt", quantity_of_tickets'.
Ticket information includes a unique code, movie title, session time,
session date, row and seat in the theater, and movie localization.

The `history_add_info(dt, active_session, movie, code);` function is
called to add information about the purchased ticket to the history file

The `generate_QR_Code` function is called to generate a QR code with a
unique ticket code and save it in the `qrFilename` file.